Fixed permissions and silencing warning

+++ shared/net/internet_layer/ipv4.h
@@ -7,7 +7,7 @@
 #ifdef __cplusplus
 extern "C" {
 #endif
-#define NET_MODE_DISABLED (-1)
+#define NET_MODE_DISABLED ((int8_t)-1)
 #define NET_MODE_DHCP 0
 #define NET_MODE_STATIC 1
